Force the size of the vt window to be a multiple of the cell size.
When the wm does not honour our size hints there are extra areas in the
vt window not covered by the terminal screen. Such gaps, when a bg
pixmap is set, would have to be cleared manually to properly refresh the
background. We take the simpler route and shrink the vt window so as to
avoid creating gaps.