Fix bug in core double-click word selection.
selection_delimit_word does not account for (left) delimiters with width > 1.
In such a case, it sets the starting position of the selection to the
second column of the delimiter (thereby including the delimiter in the
selection) rather than to the first column of the character next to it.
This bug was masked before by 'selection_make' but a recent fix to the
selection code revealed it.